use a single RulesContextMock definition
